scraping data from imdb

+++ database/migrations/2023_04_10_120539_create_movies_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->text('poster');
             $table->string('year');
             $table->string('trailer');
-            $table->integer('rating');
+            $table->float('rating');
             $table->timestamps();
         });
 
@@ -28,8 +28,6 @@ return new class extends Migration
             $table->id();
             $table->foreignId('genre_id')->constrained();
             $table->foreignId('movie_id')->constrained();
-            // $table->integer('user_rating')->nullable();
-            // $table->unique(['genre_id', 'movie_id']);
         });
     }
